Move Toggle to Toggle.js
